--- acmev2/pop_challenge.go
@@ -1,1 +1,0 @@
-package acme
