+++ providers/dns/gandiv5/gandiv5_test.go
@@ -2,13 +2,10 @@ package gandiv5
 
 import (
 	"crypto"
-	"crypto/rand"
-	"crypto/rsa"
 	"io"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"regexp"
 	"strings"
 	"testing"
@@ -89,61 +86,6 @@ func TestDNSProvider(t *testing.T) {
 	err = provider.CleanUp("abc.def.example.com", "", fakeKeyAuth)
 	if err != nil {
 		t.Fatal(err)
-	}
-}
-
-// TestDNSProviderLive performs a live test to obtain a certificate
-// using the Let's Encrypt staging server. It runs provided that both
-// the environment variables GANDIV5_API_KEY and GANDI_TEST_DOMAIN are
-// set. Otherwise the test is skipped.
-//
-// To complete this test, go test must be run with the -timeout=40m
-// flag, since the default timeout of 10m is insufficient.
-func TestDNSProviderLive(t *testing.T) {
-	apiKey := os.Getenv("GANDIV5_API_KEY")
-	domain := os.Getenv("GANDI_TEST_DOMAIN")
-	if apiKey == "" || domain == "" {
-		t.Skip("skipping live test")
-	}
-	// create a user.
-	const rsaKeySize = 2048
-	privateKey, err := rsa.GenerateKey(rand.Reader, rsaKeySize)
-	if err != nil {
-		t.Fatal(err)
-	}
-	myUser := user{
-		Email: "test@example.com",
-		key:   privateKey,
-	}
-	// create a client using staging server
-	client, err := acme.NewClient(stagingServer, &myUser, acme.RSA2048)
-	if err != nil {
-		t.Fatal(err)
-	}
-	provider, err := NewDNSProviderCredentials(apiKey)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = client.SetChallengeProvider(acme.DNS01, provider)
-	if err != nil {
-		t.Fatal(err)
-	}
-	client.ExcludeChallenges([]acme.Challenge{acme.HTTP01, acme.TLSSNI01})
-	// register and agree tos
-	reg, err := client.Register()
-	if err != nil {
-		t.Fatal(err)
-	}
-	myUser.Registration = reg
-	err = client.AgreeToTOS()
-	if err != nil {
-		t.Fatal(err)
-	}
-	// complete the challenge
-	bundle := false
-	_, failures := client.ObtainCertificate([]string{domain}, bundle, nil, false)
-	if len(failures) > 0 {
-		t.Fatal(failures)
 	}
 }
